refactor(inbox): finish the split — drop shims, repoint tests, shape seams
Final slice of the inbox.ts refactor (after #516, #517).

- Drop the temporary re-export shims from inbox.ts and repoint the 9 sibling test
  files to import directly from the new modules. inbox.ts now exports only
  `inboxRouter`.
- Shape the seams for the planned "inbox span of control" work:
  - module-map header on inbox.ts (route layer only; compose the siblings,
    don't regrow the god file),
  - labelled route bands (read / lifecycle / conversation+triage / metadata /
    actions / learnings) so where-does-a-new-endpoint-go is obvious,
  - doc-comment marking inbox-widgets.ts as the single in-thread output-widget
    boundary (the drop-in point for widgets-in-threads).

No behavior change. inbox.ts: 3217 (pre-refactor) -> 938, split across
inbox-shared (266) / inbox-catalog (395) / inbox-plan (252) / inbox-widgets (165)
/ inbox-engine (1355). Module graph strictly acyclic. Clean build; full
dashboard+core suite unchanged at 2018 pass / 3 skip.
